--- tools/maven/pom.xml
@@ -1,67 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-  ~ Copyright 2020 Red Hat, Inc., and individual contributors.
-  ~
-  ~ Licensed under the Apache License, Version 2.0 (the "License");
-  ~ you may not use this file except in compliance with the License.
-  ~ You may obtain a copy of the License at
-  ~
-  ~ http://www.apache.org/licenses/LICENSE-2.0
-  ~
-  ~ Unless required by applicable law or agreed to in writing, software
-  ~ distributed under the License is distributed on an "AS IS" BASIS,
-  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  ~ See the License for the specific language governing permissions and
-  ~ limitations under the License.
-  -->
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-  xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
-  <modelVersion>4.0.0</modelVersion>
-
-  <parent>
-    <groupId>org.wildfly-extras</groupId>
-    <artifactId>transformer-tool-parent</artifactId>
-    <version>1.0.0.Alpha1-SNAPSHOT</version>
-    <relativePath>../pom.xml</relativePath>
-  </parent>
-
-  <groupId>org.wildfly.transformer</groupId>
-  <artifactId>transformer-maven-plugin</artifactId>
-  <packaging>maven-plugin</packaging>
-  <name>maven-plugin Maven Mojo</name>
-  <url>http://maven.apache.org</url>
-  <dependencies>
-    <dependency>
-      <groupId>org.apache.maven</groupId>
-      <artifactId>maven-plugin-api</artifactId>
-      <version>2.0</version>
-    </dependency>
-    <dependency>
-      <groupId>junit</groupId>
-      <artifactId>junit</artifactId>
-      <version>3.8.1</version>
-      <scope>test</scope>
-    </dependency>
-      <dependency>
-          <groupId>org.apache.maven.plugin-tools</groupId>
-          <artifactId>maven-plugin-annotations</artifactId>
-          <version>3.6.0</version>
-      </dependency>
-    <dependency>
-      <groupId>org.apache.maven</groupId>
-      <artifactId>maven-core</artifactId>
-      <version>3.6.3</version>
-    </dependency>
-      <dependency>
-          <groupId>org.wildfly-extras</groupId>
-          <artifactId>asm-based-transformer</artifactId>
-          <version>1.0.0.Alpha1-SNAPSHOT</version>
-      </dependency>
-      <dependency>
-          <groupId>org.wildfly-extras</groupId>
-          <artifactId>transformer-shared-tool</artifactId>
-          <version>1.0.0.Alpha1-SNAPSHOT</version>
-          <scope>compile</scope>
-      </dependency>
-  </dependencies>
-</project>
