+++ tools/pom.xml
@@ -20,7 +20,6 @@
   xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
 
   <modelVersion>4.0.0</modelVersion>
-
   <parent>
     <groupId>org.wildfly-extras</groupId>
     <artifactId>batavia-parent</artifactId>
@@ -28,14 +27,37 @@
     <relativePath>../pom.xml</relativePath>
   </parent>
 
-  <name>Transformer Tools Parent</name>
-  <artifactId>transformer-tool-parent</artifactId>
-  <packaging>pom</packaging>
+  <name>Transformer Tools</name>
+  <artifactId>transformer-tools</artifactId>
+  <packaging>jar</packaging>
 
-  <modules>
-    <module>cmdline</module>
-    <module>maven</module>
-    <module>shared</module>
-  </modules>
+
+  <dependencies>
+    <dependency>
+      <groupId>org.apache.maven</groupId>
+      <artifactId>maven-plugin-api</artifactId>
+      <version>2.0</version>
+    </dependency>
+    <dependency>
+      <groupId>junit</groupId>
+      <artifactId>junit</artifactId>
+      <version>3.8.1</version>
+      <scope>test</scope>
+    </dependency>
+      <dependency>
+          <groupId>org.apache.maven.plugin-tools</groupId>
+          <artifactId>maven-plugin-annotations</artifactId>
+          <version>3.6.0</version>
+      </dependency>
+    <dependency>
+      <groupId>org.apache.maven</groupId>
+      <artifactId>maven-core</artifactId>
+      <version>3.6.3</version>
+    </dependency>
+      <dependency>
+          <groupId>org.wildfly-extras</groupId>
+          <artifactId>transformer-api</artifactId>
+      </dependency>
+  </dependencies>
 
 </project>
