+++ tools/src/main/java/org/wildfly/transformer/HandleTransformation.java
@@ -3,7 +3,7 @@ package org.wildfly.transformer;
 import java.io.File;
 import java.io.IOException;
 
-import org.wildfly.transformer.tool.shared.Common;
+import org.wildfly.transformer.transformer.tool.shared.Common;
 
 /**
  * HandleTransformation
