+++ tools/src/main/java/org/wildfly/transformer/transformer/tool/shared/Common.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.wildfly.transformer.tool.shared;
+package org.wildfly.transformer.transformer.tool.shared;
 
 import java.io.Closeable;
 import java.io.File;
